Migrated Log Prompt to new LogData file
